Fixing typo

+++ source/widynski.h
@@ -16,7 +16,7 @@ static inline void widynski_seed(uint64_t seed) {
       g_widynski_s = g_widynski_s | (g_widynski_s << 32);
 }
 
-static inline uint64_t widynski() {
+static inline uint32_t widynski() {
     g_widynski_x *= g_widynski_x;
     g_widynski_x += (g_widynski_w += g_widynski_s);
     g_widynski_x  = (g_widynski_x >> 32) | ( g_widynski_x << 32);
